Debug bool en Stop functie
Added debug bool that ensures that the HMI will not search for connections if its in debug mode, ensuring that functions can be tested while the coobt is not available. Further is a stop function made that will stop the parser whenever the stop is pressed and start when it can go from where it left off. Furtermore a few things were added to the first version of the readme.